exposure_management/requirements.txt
streamlit==1.17.1
pandas==1.3.3
psycopg2==2.9.2
opencage==1.2.2
python-dotenv==0.19.0
